+++ crates/zap-language-server/src/definitions/mod.rs
@@ -1,3 +1,5 @@
+mod namespaces;
 mod types;
 
+pub use self::namespaces::definition as definition_for_namespaces;
 pub use self::types::definition as definition_for_types;