+++ crates/zap-language-server/src/references/mod.rs
@@ -1,3 +1,5 @@
+mod namespaces;
 mod types;
 
+pub use self::namespaces::references as references_for_namespaces;
 pub use self::types::references as references_for_types;