+++ CHANGELOG.md
@@ -8,6 +8,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Added full support for namespaces, meaning they now support:
+  - Renaming
+  - Go-to definition
+  - Finding references
+
 ## `0.2.0` - June 24th, 2025
 
 ### Added